Moved the header to a separate component
