+++ src/components/Header/Header.module.scss
@@ -1,0 +1,74 @@
+@import "../../variables.scss";
+
+.header {
+  height: 10vh;
+  background-color: $black;
+  color: $white;
+  display: flex;
+  align-items: center;
+}
+.nav {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.logo {
+  & > a {
+    text-decoration: none;
+    font-size: 26px;
+    font-weight: 500;
+    color: $white;
+  }
+}
+.user {
+  display: flex;
+  align-items: center;
+
+  & > h3 {
+    margin: 0;
+  }
+  .profileImg {
+    display: flex;
+    align-items: center;
+    margin: 0 0 0 16px;
+    border-radius: 50%;
+    background: linear-gradient(-45deg, #157ae1, #c7a1ff);
+    padding: 2px;
+    width: 60px;
+    height: 60px;
+    a {
+      display: flex;
+      align-items: center;
+    }
+    &Avatar {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+      border-radius: 50%;
+      border: 4px solid $black;
+      cursor: pointer;
+    }
+  }
+  & > a {
+    text-decoration: none;
+    margin-left: 10px;
+    font-size: 16px;
+    font-weight: 500;
+    color: $blue;
+  }
+}
+
+.navbarNav {
+  display: flex;
+  .navItem {
+    list-style: none;
+    .navLink {
+      text-decoration: none;
+      margin-left: 20px;
+      font-size: 16px;
+      font-weight: 500;
+      color: $blue;
+    }
+  }
+}
